+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/MediaGroupUpdates/MediaGroupUpdate.kt
@@ -1,0 +1,8 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.update.MediaGroupUpdates
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaGroupMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
+
+interface MediaGroupUpdate: BaseMessageUpdate {
+    override val data: MediaGroupMessage
+}